build: Change internal DB port to be the same as external (#43)
Changing the DB port away from `3306` did not work because the communication between containers is within the same network, so the internal DB port was used. This made the Django app not reach the DB anymore.

To support both connecting from the outside (via the virtual environment) and from within a container, this MR changes the internal port MariaDB is listening to be the same as the port it is mapped from.

See merge request opalmedapps/opal-admin#34