docs: add page on recovering from database migration failures
Added a documentation page with information on recovering from migration failures. This change is motivated by a failure that required me to edit the database directly to fix an issue, so I'm recording what I learned in case it might be useful to someone else later.

See merge request opalmedapps/opal-admin#182